Change noise seeds from int to long
This is important on GWT, where the seeds are likely to overflow during
many operations, but an int will have very different overflow behavior
on GWT than on any normal platorm. The long type is emulated on GWT, but
it's emulated correctly, while int is anything but correct.

+++ squidlib-util/src/main/java/squidpony/squidmath/TabbyNoise.java
@@ -64,7 +64,7 @@ public static double gauss(long state) {
     }
 
     @Override
-    public double getNoiseWithSeed(final double x, final double y, final int seed) {
+    public double getNoiseWithSeed(final double x, final double y, final long seed) {
         final long
                 rs  = ThrustRNG.determine(seed ^ (long)~seed << 32),
                 rx = (rs >>> 23 ^ rs << 23) * (rs | 1L),
@@ -104,7 +104,7 @@ public static double gauss(long state) {
         return getNoiseWithSeeds(x, y, z, seedX, seedY, seedZ, randX, randY, randZ);
     }
     @Override
-    public double getNoiseWithSeed(final double x, final double y, final double z, final int seed) {
+    public double getNoiseWithSeed(final double x, final double y, final double z, final long seed) {
         final long
                 rs  = ThrustRNG.determine(seed ^ (long)~seed << 32),
                 rx = (rs >>> 23 ^ rs << 23) * (rs | 1L),
@@ -151,7 +151,7 @@ public static double gauss(long state) {
     }
 
     @Override
-    public double getNoiseWithSeed(final double x, final double y, final double z, final double w, final int seed) {
+    public double getNoiseWithSeed(final double x, final double y, final double z, final double w, final long seed) {
         final long
                 rs  = ThrustRNG.determine(seed ^ (long)~seed << 32),
                 rx = (rs >>> 23 ^ rs << 23) * (rs | 1L),
@@ -207,7 +207,7 @@ public static double gauss(long state) {
     }
 
     @Override
-    public double getNoiseWithSeed(final double x, final double y, final double z, final double w, final double u, final double v, int seed) {
+    public double getNoiseWithSeed(final double x, final double y, final double z, final double w, final double u, final double v, long seed) {
         final long
                 rs  = ThrustRNG.determine(seed ^ (long)~seed << 32),
                 rx = (rs >>> 23 ^ rs << 23) * (rs | 1L),
